Major Modularization: Refactored the codebase so that all major logic (data loading, technical indicators, prediction, PDF generation, UI components) is now organized into separate modules under the src directory. This makes the codebase much easier to maintain, extend, and test.
Combined Technical & Fundamental Metrics:
The dashboard now fetches and displays both technical indicators and fundamental metrics (EPS, P/E Ratio, Revenue Growth, Profit Margin) for each stock, providing a more comprehensive analysis.

AI-Powered Price Prediction:
Integrated a price prediction feature using scikit-learn’s RandomForestRegressor, leveraging both technical and fundamental features to predict the next day’s closing price.

Improved UI & User Experience:
The sidebar and dashboard sections are now fully modular and reusable. The sidebar supports dynamic, context-aware indicator selection and strategy configuration.

Session State & Error Handling:
Improved management of Streamlit session state for all user selections and results. Enhanced error handling and user feedback throughout the app.

PDF Reporting:
Users can generate and download comprehensive PDF reports that include charts, analysis, indicators, and risk metrics.

Updated Documentation:
The README.md has been updated to reflect the new modular structure, features, and usage instructions, making it easier for new users and contributors to get started.
