+++ README.md
@@ -2,19 +2,46 @@
 
 
 ## **Overview**
+
 This project provides an **AI-powered technical stock analysis dashboard** built with Streamlit, Plotly, and advanced technical analysis tools. The dashboard features a modern, interactive UI, dynamic strategy and indicator selection, and AI-driven market analysis with PDF reporting.
+
+**Key Features (2025 Update):**
+- Modular codebase: All major logic (data loading, prediction, PDF, UI) is in `src/` modules for maintainability.
+- Combined technical and fundamental metrics for analytics and AI prediction.
+- Price prediction using RandomForestRegressor (scikit-learn) with both technical and fundamental features.
+- Sidebar and dashboard sections are fully modular and reusable.
+- Improved session state management and error handling.
+
 
 
 
 ### **Libraries Used**
 - `streamlit`: For the interactive web dashboard
-- `yfinance`: For downloading historical stock data
+- `yfinance`: For downloading historical stock and fundamental data
 - `pandas`: For data manipulation and time series
 - `plotly`: For interactive candlestick and technical indicator charts
 - `ollama`: For sending chart images and prompts to an AI model (LLaMA 3.2 Vision)
+- `scikit-learn`: For price prediction using RandomForestRegressor
 - `tempfile`, `base64`, `os`: For temporary file management and PDF/chart handling
 
 
+
+
+## **Project Structure**
+
+```
+├── app.py/                    # Main Streamlit dashboard entry point
+├── src/
+│   ├── data_loader.py         # Data fetching utilities
+│   ├── indicators.py          # Technical indicator calculations
+│   ├── plotter.py             # Charting utilities
+│   ├── ai_analysis.py         # AI prompt and analysis logic
+│   ├── pdf_generator.py       # PDF report generation
+│   ├── data_pipeline.py       # Modular data fetching/processing pipeline
+│   ├── prediction.py          # Price prediction and fundamental metrics
+│   ├── pdf_utils.py           # PDF display/download helpers
+│   └── ui_components.py       # Sidebar, quick stats, and UI helpers
+```
 
 ## **What the Code Does**
 
@@ -45,22 +72,27 @@ This project provides an **AI-powered technical stock analysis dashboard** built
 
 
 
-### **4. Technical Indicators & Charting**
+
+### **4. Technical & Fundamental Metrics, Charting**
 - Interactive candlestick charts are created with Plotly.
 - Sidebar allows dynamic selection of technical indicators, grouped by category (momentum, trend, volatility, volume, oscillators).
 - Recommended indicators are pre-selected based on chosen strategy.
 - Indicators include: SMA, EMA, RSI, MACD, Bollinger Bands, VWAP, ATR, ADX, Stochastic, OBV, and more.
+- **Fundamental metrics** (EPS, P/E Ratio, Revenue Growth, Profit Margin) are fetched via yfinance and combined with technicals in the dashboard and for AI analysis.
 
 
 
 
-### **5. AI-Powered Analysis & PDF Reporting**
+
+### **5. AI-Powered Analysis, Price Prediction & PDF Reporting**
 - Uses the `ollama` library (LLaMA 3.2 Vision) for advanced chart and indicator analysis.
 - When users click "Run Analysis 💸":
   - The chart is saved as an image and sent to the AI model with a detailed, context-aware prompt.
   - The AI model provides a trade recommendation, strategy, and rationale.
+- **Price prediction**: The dashboard predicts the next day's closing price using a RandomForestRegressor trained on both technical and fundamental features.
 - Results are displayed in the dashboard.
 - Users can generate a comprehensive PDF report (with chart, analysis, indicators, and risk metrics) for preview and download.
+
 
 
 
@@ -68,9 +100,17 @@ This project provides an **AI-powered technical stock analysis dashboard** built
 ### **6. Temporary File & State Management**
 - Temporary files for charts and PDFs are managed and cleaned up automatically.
 - All user selections and results are managed via Streamlit session state for a seamless experience.
+- Modular session state keys for stock data, levels, options data, and active indicators.
 
 
 
+
+
+## **Modularization & Maintainability**
+
+- All major logic is separated into modules in `src/` for easy maintenance and extension.
+- UI components (sidebar, quick stats) are reusable and easy to update.
+- Data pipeline and prediction logic are decoupled from the main app for clarity.
 
 ## **Use Cases**
 
@@ -119,7 +159,7 @@ conda deactivate <ENV_NAME>
    - Keep this terminal open while using the dashboard.
 2. In a new terminal, install the required libraries:
    ```bash
-   pip install streamlit plotly ollama pandas yfinance
+   pip install streamlit plotly ollama pandas yfinance scikit-learn
    ```
 3. Run the app:
    ```bash
